adding product

+++ src/main/java/com/vicheak/phoneshop/project/controller/BrandController.java
@@ -47,13 +47,13 @@ public class BrandController {
 	}
 	
 	@GetMapping("{id}")
-	public ResponseEntity<?> getOneBrand(@PathVariable("id") Integer brandId){
+	public ResponseEntity<?> getOneBrand(@PathVariable("id") Long brandId){
 		Brand brand = brandService.getById(brandId);
 		return ResponseEntity.ok(BrandMapper.INSTANCE.toBrandDTO(brand)); 
 	}
 	
 	@PutMapping("{id}")
-	public ResponseEntity<?> updateBrand(@PathVariable("id") Integer brandId,
+	public ResponseEntity<?> updateBrand(@PathVariable("id") Long brandId,
 										 @RequestBody BrandDTO brandDTO){
 		//Brand brand = Mapper.toBrand(brandDTO);
 		Brand brand = BrandMapper.INSTANCE.toBrand(brandDTO);
@@ -92,7 +92,7 @@ public class BrandController {
 	}*/
 	
 	@GetMapping("{id}/models")
-	public ResponseEntity<?> getModelsByBrand(@PathVariable("id") Integer brandId){
+	public ResponseEntity<?> getModelsByBrand(@PathVariable("id") Long brandId){
 		List<Model> models = modelService.getByBrand(brandId);
 		List<ModelDTO> list = models.stream()
 			.map(modelMapper::toModelDTO)
