+++ src/main/java/com/vicheak/phoneshop/project/repository/ModelRepository.java
@@ -5,10 +5,9 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.vicheak.phoneshop.project.entity.Brand;
 import com.vicheak.phoneshop.project.entity.Model;
 
 @Repository
-public interface ModelRepository extends JpaRepository<Model, Integer>{
-	List<Model> findByBrandId(Integer brandId); 
+public interface ModelRepository extends JpaRepository<Model, Long>{
+	List<Model> findByBrandId(Long brandId); 
 }
